backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
redis==5.0.8
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
PyJWT==2.9.0
